ipsk manager database add create endpoint

+++ supportfiles/portals/sponsorportal/default/delete.inc.php
@@ -37,16 +37,33 @@
 		
 		if($endpointPermissions){
 			if($endpointPermissions[0]['advancedPermissions'] & 64){
-		
-				$endPointAssociation = $ipskISEDB->getEndPointAssociationById($sanitizedInput['id']);
-				
 				//LOG::Entry
 				$logData = $ipskISEDB->generateLogData(Array("sanitizedInput"=>$sanitizedInput));
 				$logMessage = "REQUEST:SUCCESS;ACTION:SPONSORDELETE;METHOD:DELETE-ENDPOINT;MAC:".$sanitizedInput['macAddress'].";REMOTE-IP:".$_SERVER['REMOTE_ADDR'].";USERNAME:".$_SESSION['logonUsername'].";SID:".$_SESSION['logonSID'].";";
 				$ipskISEDB->addLogEntry($logMessage, __FILE__, __FUNCTION__, __CLASS__, __METHOD__, __LINE__, $logData);
-				
-				$ipskISEDB->deleteEndpointAssociationbyId($sanitizedInput['id']);
-				$ipskISEDB->deleteEndpointById($endPointAssociation['endpointId']);
+				$myEndPoint = $ipskISEDB->getEndpointById($endPointAssociation['endpointId']);
+				$ersCreds = $ipskISEDB->getISEERSSettings();
+				$logData = $ipskISEDB->generateLogData(Array("sanitizedInput"=>$sanitizedInput));
+				$logMessage = "DELETE:DEBUG;ACTION:SPONSOREDIT;METHOD:DELETE-ENDPOINT-ISE;SID:".$_SESSION['logonSID'].";";
+				$ipskISEDB->addLogEntry($logMessage, __FILE__, __FUNCTION__, __CLASS__, __METHOD__, __LINE__, $logData);
+				if($ersCreds['enabled'])
+				{
+					if($ipskISEERS->DeleteEndPointByMac($myEndPoint['macAddress'])){
+						//LOG::Entry
+						$logData = $ipskISEDB->generateLogData(Array("sanitizedInput"=>$sanitizedInput));
+						$logMessage = "REQUEST:SUCCESS;ACTION:SPONSORDELETE;METHOD:DELETE-ENDPOINT-ISE;MAC:".$myEndPoint['macAddress'].";REMOTE-IP:".$_SERVER['REMOTE_ADDR'].";USERNAME:".$_SESSION['logonUsername'].";SID:".$_SESSION['logonSID'].";";
+						$ipskISEDB->addLogEntry($logMessage, __FILE__, __FUNCTION__, __CLASS__, __METHOD__, __LINE__, $logData);
+						//delete from database if ISE delete is succesful
+						$endPointAssociation = $ipskISEDB->getEndPointAssociationById($sanitizedInput['id']);
+						$ipskISEDB->deleteEndpointAssociationbyId($sanitizedInput['id']);
+						$ipskISEDB->deleteEndpointById($endPointAssociation['endpointId']);
+					}else{
+						//LOG::Entry
+						$logData = $ipskISEDB->generateLogData(Array("sanitizedInput"=>$sanitizedInput));
+						$logMessage = "REQUEST:FAILURE;ACTION:SPONSORDELETE;METHOD:DELETE-ENDPOINT-ISE;MAC:".$myEndPoint['macAddress'].";REMOTE-IP:".$_SERVER['REMOTE_ADDR'].";USERNAME:".$_SESSION['logonUsername'].";SID:".$_SESSION['logonSID'].";";
+						$ipskISEDB->addLogEntry($logMessage, __FILE__, __FUNCTION__, __CLASS__, __METHOD__, __LINE__, $logData);
+					}
+				}
 
 				print <<<HTML
 <script>
@@ -64,7 +81,7 @@ HTML;
 			<div class="modal-header shadow alert alert-danger">
 				<h5 class="modal-title font-weight-bold" id="modalLongTitle">Delete Endpoint Association?</h5>
 				<button type="button" class="close" data-dismiss="modal" aria-label="Close">
-				  <span aria-hidden="true">&times;</span>
+				<span aria-hidden="true">&times;</span>
 				</button>
 			</div>
 			<div class="modal-body">
